x509: various test cleanups
